Reformatting and small bug fixes
- UB fix: __builtin_clzll(0) in CRC validation — added sum != 0 guard
- Wrong topic: Sensor timeout published to {serial}/status instead of sensor/{serial}/status
- Global keyfob dedup: Single last_keyfob_payload suppressed duplicate payloads across different keyfobs — now per-serial map
- Static local in member function: last_diag_time was static, shared across instances — moved to member variable
- Double slash in topics: Both main.cpp and DigitalDecoder constructor appended / to topic prefix — removed from main.cpp
- Float/double mismatch: FILTER_ALPHA was double causing implicit promotion/narrowing on every RF sample — changed to float

+++ src/analog_decoder.cpp
@@ -7,10 +7,10 @@ constexpr int HW_RATIO = 17;
 constexpr float MIN_OOK_THRESHOLD = 0.25f;
 constexpr float OOK_THRESHOLD_RATIO = 0.75f;
 constexpr float OOK_DECAY_PER_SAMPLE = 0.0001f;
-constexpr double FILTER_ALPHA = 0.7;
+constexpr float FILTER_ALPHA = 0.7f;
 
 void AnalogDecoder::handle_magnitude(float value) {
-    val = FILTER_ALPHA * val + (1.0 - FILTER_ALPHA) * value;
+    val = FILTER_ALPHA * val + (1.0f - FILTER_ALPHA) * value;
     value = val;
 
     if (discarded_samples < (HW_RATIO - 1)) {
